+++ modules/access/mms/mmstu.c
@@ -68,7 +68,7 @@ int  E_( MMSTUOpen )  ( access_t * );
 void E_( MMSTUClose ) ( access_t * );
 
 
-static int Read( access_t *, uint8_t *, int );
+static ssize_t Read( access_t *, uint8_t *, size_t );
 static int Seek( access_t *, int64_t );
 static int Control( access_t *, int, va_list );
 
@@ -370,7 +370,7 @@ static int Seek( access_t * p_access, int64_t i_pos )
 /*****************************************************************************
  * Read:
  *****************************************************************************/
-static int Read( access_t *p_access, uint8_t *p_buffer, int i_len )
+static ssize_t Read( access_t *p_access, uint8_t *p_buffer, size_t i_len )
 {
     access_sys_t *p_sys = p_access->p_sys;
     size_t      i_data;
